Debugging
Added two commandline parameters. --selectionmode has two options for
now, random and first. --debug has three options which give varying
degrees of debugging printouts: none, some and all.
